Bug 8773 - Start per-instance koha-index-daemon in .deb setup
Short:

Launch a koha-index-daemon (from Koha::Contrib::tamil) process for each
enabled instance. Enabling/disabling the use of the indexer is handled
using a new(ly introduced) /etc/default/koha-common config file that
can be used for other stuff too, and is the Debian way of handling
init scripts control variables[1].

Also provide command line tools to manage the running indexer daemons
for your instances.

Long:

Using an indexing daemon avoids launching a new interpreter each time
the cron triggers the indexing, and also allows sub-minute incremental
reindexing, a requirement from our librarians.[2]

Using the indexer daemon should remain "experimental" so is disabled by
default. To enable the use of the indexer the user has to tweak the
/etc/default/koha-common config file (introduced by this patch). Specifically
the USE_INDEXER_DAEMON variable, which is clearly explained in the file.

Update frecquency defaults to 30 sec, and can be changed by tweaking the
/etc/default/koha-common config file too.

This patch relies on Koha::Contrib::Tamil's koha-index-daemon script [3],
but could be easily adjusted to use another if we got a new one (SolR?).
As Robin proposed it checks for availability of the relevant files before
trying to spawn the daemon.

If there's something I could change to make this patches better just let me know.

Note: there was a problem Robin found regarding the spawned processes forgetting
the PERL5LIB and/or KOHA_CONF env variables, this is fixed in this version
with the inclusion of the 'export' keyword for the relevant variables.

Edit: fixed koha-create to call the koha-indexer script instead, and added
-q to make koha-inxeder quiet (useful for the init script).

Regards
To+

[1] Section 9.3.2 of http://www.debian.org/doc/debian-policy/ch-opersys.html
[2] This is the .deb version of http://bugs.koha-community.org/bugzilla3/show_bug.cgi?id=8519
[3] The default cronjob that is set using the debian/koha-common.cron.d file
is being disbaled by this patch to avoid collision.

Sponsored-by: Universidad Nacional de Cordoba