Koha::Filter::MARCXML::ViewPolicy introduced
This patch introduces a Koha::RecordProcessor filter for MARCXML data.
Specifically, it implements the visibility policy using the frameworks
to do the job. It also replaces authorized value codes for descriptions
using the framework's subfield->authorized value category configuration.

It does the job by creating an XSLT stream to be applied to the passed
XML::LibXML object.

The key ideas on this implementation are:

- Faster processing (several stuff done traversing the record only once).
- The framework's information is read only once, and cached for reuse!
- Code is MUCH simpler and compact (modulo XSLT's inherent complexity).

To be continued