Koha::Search::Result introduced
This patch introduces a Koha::Search::Result class to be used for building
search results. The idea is that each SearchEngine implementation should
return this objects, or a list of them.

This objects are intended to be used by Koha::RecordProcessor to implement
the visibility policy (framework visibility, authorized value descriptions,
etc).