Bug 13791: (QA followup) Plack support requires Apache 2.4.8+
The Apache 2.4 documentation says UDS support for ProxyPass has been
introduced in Apache 2.4.7. That's what is shipped by Koha's supported
Debian / Ubuntu versions.
But it is not true, empiric tests and even the changelog for the apache
project says the feature got introduced in 2.4.8. So to avoid breakages
I raised the version number test on the apache files to 2.4.8.

New bugs will be filled for dealing with this if we find it necessary.

Signed-off-by: Tomas Cohen Arazi <tomascohen@unc.edu.ar>