Bug 16016: Apache file tweak
This patch makes apache expose the sitemap files. This is suitable for
including the sitemap in the robots.txt file as proposed in

  http://www.sitemaps.org/protocol.html#submit_robots

Note: it depends on Apache 2.4+ so we can patch the file as needed. Older
Apache users will be warned that they need to tweak the apache files on their
own.

Sponsored-by: Orex Digital