Bug 13098: Tests for _FixAccountForLostAndReturned on writeoff
When a a write off is done on lost fees, but then the item is returned,
the current behaviour is that a credit is added to the patron, for the
amount the original lost fee was. This behaviour is wrong and this tests
reproduce the situation.

To test:
- Run:
  $ kshell
 k$ prove t/db_dependent/Circulation.t
=> FAIL: Tests fail because a credit (CR) is added for the original
amount, without taking into account the fact the amount was not payed,
but written off